+++ composer.json
@@ -48,7 +48,7 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Dot\\Tests\\": "tests/"
+            "DotTest\\ErrorHandler\\": "test/"
         }
     },
     "minimum-stability": "stable",
